requirements.txt

pip install
 fastapi
  uvicorn
   sqlalchemy
    pymysql
     passlib[bcrypt]
      python-jose
       transformers
        torch
         safetensors
          python-multipart
           pydantic[email]
            PyJWT
             scikit-learn
              pandas
               matplotlibfrom 
                bcrypt
                 python-dotenv
                  pytest httpx
                